Fixed #7 and a bug where a disabled switch would still move
+ Fixed handler width bug
+ Fixed disabled switch behaviour
+ Updated example
+ Example now works also for Android
+ Updated .npmignore and added .gitignore for future collaborations
