+++ .npmignore
@@ -1,1 +1,5 @@
 Example/
+.gitignore
+.idea
+*.iml
+*.log
