+++ lib/index.js
@@ -1,12 +1,5 @@
-import React, { Component, PropTypes } from 'react'
-import {
-  View,
-  StyleSheet,
-  Animated,
-  Easing,
-  PanResponder,
-  TouchableOpacity
-} from 'react-native'
+import React, {Component, PropTypes} from "react";
+import {Animated, Easing, PanResponder, StyleSheet, View} from "react-native";
 
 export default class extends Component {
   static propTypes = {
@@ -78,14 +71,15 @@ export default class extends Component {
   }
 
   _onPanResponderGrant = (evt, gestureState) => {
-    const { height } = this.props
-
+    const { height, disabled } = this.props
+    if (disabled) return;
     this.animateHandler(height * 6 / 5)
   }
 
   _onPanResponderMove = (evt, gestureState) => {
     const { value, toggleable } = this.state
-
+    const { disabled} = this.props
+    if (disabled) return;
     this.setState({
       toggleable: value ? (gestureState.dx < 10) : (gestureState.dx > -10)
     })
@@ -94,13 +88,15 @@ export default class extends Component {
   _onPanResponderRelease = (evt, gestureState) => {
     const { handlerAnimation, toggleable, value } = this.state
     const { height, disabled, onAsyncPress, onSyncPress } = this.props
-
-    if (toggleable && !disabled) {
+    if (disabled) return;
+    if (toggleable) {
       if (onSyncPress) {
         this.toggleSwitch(true, onSyncPress)
       } else {
         onAsyncPress(this.toggleSwitch)
       }
+    } else {
+      this.animateHandler(this.handlerSize)
     }
   }
 
@@ -139,7 +135,7 @@ export default class extends Component {
     Animated.timing(handlerAnimation,
       {
         toValue: value,
-        duration: 200,
+        duration: 50,
         easing: Easing.linear
       }
     ).start(callback)
@@ -157,7 +153,7 @@ export default class extends Component {
       inputRange: value ? [-this.offset, -1]: [1, this.offset],
       outputRange: [backgroundInactive, backgroundActive]
     })
-    
+
     const interpolatedCircleColor = switchAnimation.interpolate({
       inputRange: value ? [-this.offset, -1]: [1, this.offset],
       outputRange: [circleColorInactive, circleColorActive]
@@ -168,17 +164,17 @@ export default class extends Component {
         {...rest}
         {...this._panResponder.panHandlers}
         style={[styles.container, style, {
-        width, height,
-        alignItems,
-        borderRadius: height / 2,
-        backgroundColor: interpolatedBackgroundColor }]}>
+          width, height,
+          alignItems,
+          borderRadius: height / 2,
+          backgroundColor: interpolatedBackgroundColor }]}>
         <Animated.View style={{
           backgroundColor: interpolatedCircleColor,
           width: handlerAnimation,
           height: this.handlerSize,
           borderRadius: height / 2,
           transform: [{ translateX: switchAnimation }]
-          }} />
+        }} />
       </Animated.View>
     )
   }
